requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
importlib_resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
packaging==24.1
psycopg2==2.9.9
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
swagger-spec-validator==3.0.4
transbank-sdk==5.0.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
